requirements.txt
asgiref==3.6.0
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.4.0
